requirements.txt
flask
requests
pdfminer.six
python-docx
flask-cors
transformers